removed helpers module. Added dags fast tests. Updated typing and description for print_acc_by_covars to specify usage with ContextualizedClassifier
