+++ tests.py
@@ -8,6 +8,7 @@ from contextualized.easy.tests.test_markov_networks import *
 from contextualized.easy.tests.test_correlation_networks import *
 from contextualized.easy.tests.test_bayesian_networks import *
 from contextualized.analysis.tests import *
+from contextualized.baselines.tests import *
 
 if __name__ == '__main__':
     unittest.main()
